* Finished rebase adaptations

+++ pallets/battle-mogs/src/tests.rs
@@ -526,13 +526,13 @@ mod sacrifice_into {
 			// We need to up the rarity in order to be allowed to sacrifice
 			Mogwais::<Test>::mutate(mogwai_id_1, |maybe_mogwai| {
 				if let Some(ref mut mogwai) = maybe_mogwai {
-					mogwai.rarity = RarityType::Epic as u8;
+					mogwai.rarity = RarityType::Epic;
 				}
 			});
 
 			Mogwais::<Test>::mutate(mogwai_id_2, |maybe_mogwai| {
 				if let Some(ref mut mogwai) = maybe_mogwai {
-					mogwai.rarity = RarityType::Epic as u8;
+					mogwai.rarity = RarityType::Epic;
 				}
 			});
 
@@ -575,7 +575,7 @@ mod sacrifice_into {
 
 			Mogwais::<Test>::mutate(mogwai_id_1, |maybe_mogwai| {
 				if let Some(ref mut mogwai) = maybe_mogwai {
-					mogwai.rarity = RarityType::Common as u8;
+					mogwai.rarity = RarityType::Common;
 				}
 			});
 
@@ -596,13 +596,13 @@ mod sacrifice_into {
 			// We need to up the rarity in order to be allowed to sacrifice
 			Mogwais::<Test>::mutate(mogwai_id_1, |maybe_mogwai| {
 				if let Some(ref mut mogwai) = maybe_mogwai {
-					mogwai.rarity = RarityType::Epic as u8;
+					mogwai.rarity = RarityType::Epic;
 				}
 			});
 
 			Mogwais::<Test>::mutate(mogwai_id_2, |maybe_mogwai| {
 				if let Some(ref mut mogwai) = maybe_mogwai {
-					mogwai.rarity = RarityType::Epic as u8;
+					mogwai.rarity = RarityType::Epic;
 				}
 			});
 
@@ -623,13 +623,13 @@ mod sacrifice_into {
 
 			Mogwais::<Test>::mutate(mogwai_id_1, |maybe_mogwai| {
 				if let Some(ref mut mogwai) = maybe_mogwai {
-					mogwai.rarity = RarityType::Epic as u8;
+					mogwai.rarity = RarityType::Epic;
 				}
 			});
 
 			Mogwais::<Test>::mutate(mogwai_id_2, |maybe_mogwai| {
 				if let Some(ref mut mogwai) = maybe_mogwai {
-					mogwai.rarity = RarityType::Epic as u8;
+					mogwai.rarity = RarityType::Epic;
 				}
 			});
 
@@ -648,7 +648,7 @@ mod sacrifice_into {
 
 			Mogwais::<Test>::mutate(mogwai_id_1, |maybe_mogwai| {
 				if let Some(ref mut mogwai) = maybe_mogwai {
-					mogwai.rarity = RarityType::Epic as u8;
+					mogwai.rarity = RarityType::Epic;
 				}
 			});
